src/requirements.txt
flask
flask-cors
python-keycloak
python-jose
python-dotenv
werkzeug
